refactor(triggers): split triggers.ts into supervisor modules
triggers.ts was 876 lines of trigger-supervisor logic over shared
module-level registries. Split into a triggers/ folder:
  - state   — injected deps + child/timeout/log/buffer registries, warden
              set, timing constants, init + getRunningCount
  - command — interpreter resolution per script language
  - output  — stdout/stderr capture, payload truncation, wake firing
  - exit    — timeout/cancel/shutdown, child kill, finalizeExit, failTrigger
  - spawn   — the warden + direct spawn paths
  - resume  — post-restart respawn + orphan probe

index.ts re-exports the public surface AND the `_internals` test hook
(pointing at the same shared Maps). All importers updated to the barrel
(core/tools + gateway-actions own ./triggers.js untouched). Verified:
triggers + triggers-extended + lua-runner suites pass (59 tests).
