+++ src/__tests__/triggers-extended.test.ts
@@ -44,7 +44,7 @@ const {
   resumeAfterRestart,
   shutdownTriggers,
   _internals,
-} = await import("../core/background/triggers.js");
+} = await import("../core/background/triggers/index.js");
 
 import type { Trigger } from "../storage/trigger-store.js";
 const {
